movies_recommendation/requirements_api.txt
#gradio>=4.2.0
annoy>=1.16.3
flask>=2.3.2
pillow>=9.4.0
#requests>=2.31
matplotlib>=3.7.1
numpy>=1.23.5
torch>= 2.1.1
torchvision>=0.16.1
gdown>=4.6.6
